+++ tests/SlmLocaleTest/Locale/DetectorTest.php
@@ -39,21 +39,20 @@
  */
 namespace SlmLocaleTest\Locale;
 
-use PHPUnit_Framework_TestCase as TestCase;
+use Laminas\EventManager\EventManager;
+use Laminas\Stdlib\Request;
+use Laminas\Stdlib\Response;
+use PHPUnit\Framework\TestCase;
 use SlmLocale\Locale\Detector;
 use SlmLocale\LocaleEvent;
-
 use SlmLocale\Strategy\AbstractStrategy;
 use SlmLocale\Strategy\StrategyInterface;
-use Zend\EventManager\EventManager;
-use Zend\Stdlib\Request;
-use Zend\Stdlib\Response;
 
 class DetectorTest extends TestCase
 {
     private $events;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->events = new EventManager();
     }
@@ -202,8 +201,8 @@ class DetectorTest extends TestCase
     public function testStrategyAttachesToEventManager()
     {
         $detector = new Detector();
-        $events   = $this->getMock(EventManager::class);
-        $strategy = $this->getMock(StrategyInterface::class);
+        $events   = $this->createMock(EventManager::class);
+        $strategy = $this->createMock(StrategyInterface::class);
         $strategy->expects($this->once())
             ->method('attach')
             ->with($events);
@@ -217,12 +216,18 @@ class DetectorTest extends TestCase
         $detector  = new Detector();
         $this->setEventManager($detector);
 
-        $strategy1 = $this->getMock(AbstractStrategy::class, ['detect']);
+        $strategy1 = $this
+            ->getMockBuilder(AbstractStrategy::class)
+            ->setMethods(['detect'])
+            ->getMock();
         $strategy1->expects($this->once())
                   ->method('detect')
                   ->will($this->returnValue('Foo'));
 
-        $strategy2 = $this->getMock(AbstractStrategy::class, ['detect']);
+        $strategy2 = $this
+            ->getMockBuilder(AbstractStrategy::class)
+            ->setMethods(['detect'])
+            ->getMock();
         $strategy2->expects($this->never())
                   ->method('detect');