+++ tests/SlmLocaleTest/Strategy/HostStrategyTest.php
@@ -2,13 +2,14 @@
 
 namespace SlmLocaleTest\Strategy;
 
-use PHPUnit_Framework_TestCase as TestCase;
+use Laminas\Http\PhpEnvironment\Request;
+use Laminas\Stdlib\Parameters;
+use Laminas\Stdlib\RequestInterface;
+use Laminas\Uri\Uri;
+use PHPUnit\Framework\TestCase;
 use SlmLocale\LocaleEvent;
+use SlmLocale\Strategy\Exception\InvalidArgumentException;
 use SlmLocale\Strategy\HostStrategy;
-use Zend\Http\PhpEnvironment\Request;
-use Zend\Stdlib\Parameters;
-use Zend\Stdlib\RequestInterface;
-use Zend\Uri\Uri;
 
 class HostStrategyTest extends TestCase
 {
@@ -24,24 +25,23 @@ class HostStrategyTest extends TestCase
     public function testDetectWithoutSupportedReturnsNull()
     {
         $event = new LocaleEvent();
-        $event->setRequest($this->getMockForAbstractClass(\Zend\Http\Request::class));
+        $event->setRequest($this->getMockForAbstractClass(\Laminas\Http\Request::class));
         $event->setSupported([]);
 
         $strategy = new HostStrategy();
         $this->assertNull($strategy->detect($event));
     }
 
-    /**
-     * @expectedException \SlmLocale\Strategy\Exception\InvalidArgumentException
-     */
     public function testDetectWithoutDomainThrowsInvalidArgumentException()
     {
         $event = new LocaleEvent();
-        $event->setRequest($this->getMockForAbstractClass(\Zend\Http\Request::class));
+        $event->setRequest($this->getMockForAbstractClass(\Laminas\Http\Request::class));
         $event->setSupported(['en_GB', 'de_DE']);
 
         $strategy = new HostStrategy();
         $strategy->setOptions(['domain' => 'test']);
+
+        $this->expectException(InvalidArgumentException::class);
         $this->assertNull($strategy->detect($event));
     }