+++ tests/SlmLocaleTest/Strategy/QueryStrategyTest.php
@@ -39,14 +39,12 @@
  */
 namespace SlmLocaleTest\Locale;
 
-use PHPUnit_Framework_TestCase as TestCase;
-
+use Laminas\Http\Request  as HttpRequest;
+use Laminas\Http\Response as HttpResponse;
+use Laminas\Uri\Uri;
+use PHPUnit\Framework\TestCase;
 use SlmLocale\LocaleEvent;
 use SlmLocale\Strategy\QueryStrategy;
-
-use Zend\Http\Request  as HttpRequest;
-use Zend\Http\Response as HttpResponse;
-use Zend\Uri\Uri;
 
 class QueryStrategyTest extends TestCase
 {
@@ -59,7 +57,7 @@ class QueryStrategyTest extends TestCase
      */
     protected $event;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->strategy = new QueryStrategy();
         $this->event    = new LocaleEvent();